refactor(backend): split reports router god-file into data + service modules
reports.py mixed three concerns in the HTTP layer. Extract them, leaving the
router as thin handlers + the snapshot orchestration (which legitimately
composes the portfolio router, so it stays to avoid a service->router cycle):

- constants/report_package.py      <- static CONTRACT/NOTES/TRACEABILITY dicts
- services/report_traceability.py  <- source/ledger anchor assembly + payload
- services/annualized_income.py    <- annualized-income schedule generation
- services/report_package.py       <- pure snapshot serialize/map/CSV helpers

reports.py: 2244 -> 936 lines. Pure structural refactor, no behavior change:
routes keep paths/signatures, the names tests patch stay imported in the
router, and the moved anchor helpers are repointed in the one test that
imports them. Total LOC is ~flat (per-module overhead); the win is layering
and testability.

ruff clean; router imports with all 25 routes; reporting + package-contract +
supervisor suites (283 tests) green.
